routes all completed and functional

+++ controllers/userController.js
@@ -1,4 +1,4 @@
-const {User} = require("../models")
+const {User, Thought} = require("../models")
 
 module.exports = {
   async getAllUsers(req, res) {
@@ -40,7 +40,7 @@ module.exports = {
 
   async updateUser(req, res) {
     try {
-      const course = await User.findOneAndUpdate(
+      const user = await User.findOneAndUpdate(
         {_id: req.params.userId},
         {$set: req.body},
         {runValidators: true, new: true}
@@ -58,14 +58,43 @@ module.exports = {
 
   async deleteUser(req, res) {
     try {
-      const user = await User.findOneAndDelete(
-        {_id: req.params.courseId},
-        {$pull: {thought: req.params.thoughtId}}
-      )
+      const user = await User.findOneAndDelete({_id: req.params.userId})
 
       if (!user) {
         res.status(404).json({message: "No user found"})
       }
+      await Thought.deleteMany({
+        _id: {
+          $in: user.thoughts,
+        },
+      })
+      res.json({message: "succesfully deleted user"})
+    } catch (err) {
+      res.status(500).json(err)
+    }
+  },
+
+  async addNewFriend(req, res) {
+    try {
+      const user = await User.findOneAndUpdate(
+        {_id: req.params.userId},
+        {$addToSet: {friends: req.params.friendId}},
+        {runValidators: true, new: true}
+      )
+      res.json({message: "Friend added!"})
+    } catch (err) {
+      res.status(500).json(err)
+    }
+  },
+
+  async removeFriend(req, res) {
+    try {
+      const user = await User.findOneAndUpdate(
+        {_id: req.params.userId},
+        {$pull: {friends: req.params.friendId}},
+        {runValidators: true, new: true}
+      )
+      res.json({message: "Friend Removed!"})
     } catch (err) {
       res.status(500).json(err)
     }
